+++ src/components/molecules/Modal/index.ts
@@ -1,1 +1,2 @@
 export { default as Modal } from './Modal';
+export { default as CareerContentModal } from './CareerContentModal';
